src/requirements.txt
asgiref==3.2.3
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.8
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-developer-panel==0.1.2
djangorestframework==3.10.3
djangorestframework-api-key==1.4.1
drf-yasg==1.17.0
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-speech==1.3.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jsonfield==2.0.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==19.2
Pillow==6.2.1
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydub==0.23.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.7
